move optimism node-builder impls to new file
