+++ crates/node-builder/src/evm.rs
@@ -9,12 +9,6 @@ use reth_primitives::{
 #[derive(Debug)]
 pub struct EthEvmConfig;
 
-// TODO: remove after implementing
-// #[cfg(feature = "optimism")]
-/// Optimism-related EVM configuration.
-#[derive(Debug)]
-pub struct OptimismEvmConfig;
-
 impl EvmEnvConfig for EthEvmConfig {
     fn fill_tx_env<T>(tx_env: &mut TxEnv, transaction: T, sender: Address)
     where
@@ -32,21 +26,3 @@ impl EvmEnvConfig for EthEvmConfig {
         fill_cfg_env(cfg_env, chain_spec, header, total_difficulty)
     }
 }
-
-// TODO:
-// * split up fill_tx_env into fill_tx_env and fill_tx_env_optimism
-// * code duplication for op
-// * make fill_tx_env_optimism accept regular fill_tx_env args
-// * do the following code inside trait impl
-
-// #[cfg(feature = "optimism")]
-// {
-//     let mut envelope_buf = Vec::with_capacity(transaction.length_without_header());
-//     transaction.encode_enveloped(&mut envelope_buf);
-//     fill_tx_env(&mut self.evm.env.tx, transaction, sender, envelope_buf.into());
-// }
-
-// this is default eth fill_tx_env
-// // Fill revm structure.
-// #[cfg(not(feature = "optimism"))]
-// fill_tx_env(&mut self.evm.env.tx, transaction, sender);
