+++ crates/node-builder/src/lib.rs
@@ -12,13 +12,13 @@
 pub mod engine;
 pub use engine::EthEngineTypes;
 
-#[cfg(feature = "optimism")]
-pub use engine::OptimismEngineTypes;
-
 /// Exports commonly used concrete instances of the [EvmEnvConfig](reth_node_api::EvmEnvConfig)
 /// trait.
 pub mod evm;
 pub use evm::EthEvmConfig;
 
+/// Exports optimism-specific types that implement traits in [reth_node_api].
 #[cfg(feature = "optimism")]
-pub use evm::OptimismEvmConfig;
+pub mod optimism;
+#[cfg(feature = "optimism")]
+pub use optimism::{OptimismEngineTypes, OptimismEvmConfig};
